settings effect creator viewport and move most into shared <3
