--- Polytoria/scripts/client/settings/GraphicsPresetManager.cs.uid
@@ -1,1 +1,0 @@
-uid://c7q0dlcgmde0p
