+++ Polytoria/scripts/creator/ui/tabs/world/WorldContainer.cs
@@ -3,6 +3,8 @@
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
 using Godot;
+using Polytoria.Client.Settings.Appliers;
+using Polytoria.Creator.Settings;
 using Polytoria.Creator.UI.Gizmos;
 using Polytoria.Datamodel;
 using Polytoria.Shared;
@@ -48,6 +50,16 @@ public sealed partial class WorldContainer : SubViewportContainer
 
 	public override void _Ready()
 	{
+		if (CreatorSettingsService.Instance != null)
+		{
+			var applier = CreatorSettingsService.Instance.GetNodeOrNull<GraphicsSettingsApplier>(GraphicsSettingsApplier.NodeName);
+			if (applier != null)
+			{
+				applier.RenderViewport = _subViewport;
+				applier.ApplyViewportSettings();
+			}
+		}
+
 		// Call on next frame to actually grab focus
 		Callable.From(GrabFocus).CallDeferred();
 		base._Ready();
